backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vigen-backend"
version = "0.1.0"
description = "VIGEN Phase 1 API and worker foundation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "sqlalchemy>=2.0.36",
  "alembic>=1.14.0",
  "pydantic>=2.10.0",
  "pydantic-settings>=2.6.1",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "email-validator>=2.2.0",
  "aiosqlite>=0.20.0",
  "asyncpg>=0.30.0",
  "celery[redis]>=5.4.0",
  "redis>=5.2.0",
  "structlog>=24.4.0",
  "httpx>=0.28.0",
  "python-multipart>=0.0.17",
  "boto3>=1.35.0",
  "litellm>=1.52.0",
  "librosa>=0.10.2",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

